Expression now works on top of Bigdecimal

+++ src/main/scala/sk/drunkenpanda/bot/plugins/calc/Calculator.scala
@@ -2,7 +2,7 @@ package sk.drunkenpanda.bot.plugins.calc
 
 class Calculator {
 
-  def evaluate(exp: Expression): Double = exp match {
+  def evaluate(exp: Expression): BigDecimal = exp match {
     case Number(value) => value
     case UnaryOperator("-", exp) => -evaluate(exp)
     case BinaryOperator("+", left, right) => evaluate(left) + evaluate(right)
